pyproject.toml
[tool.poetry]
name = "zhdiary-web"
version = "0.4.0"
description = "Customizable chinese flashcards on the web"
authors = ["patarapolw <patarapolw@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
pymongo = "^3.8"
wordfreq = "^2.2"
regex = "=2018.02.21"
dnspython = "^1.16"
python-dotenv = "^0.10.1"
jieba = "^0.39.0"
flask = "^1.0"
flask-pymongo = "^2.3"
passlib = "^1.7"
Flask-HTTPAuth = "^3.3"
flask-cors = "^3.0"
gunicorn = "^19.9"
authlib = "^0.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
